foundation_project/requirements.txt
torch
torchvision
transformers>=4.53.0
accelerate
pillow
opencv-python
matplotlib
numpy
huggingface_hub
safetensors
